# openapi/opencode.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "opencode",
    "description": "opencode api",
    "version": "0.0.3"
  },
  "paths": {
...
    "/global/config": {
      "get": {
        "operationId": "global.config.get",
        "summary": "Get global configuration",
        "description": "Retrieve the current global OpenCode configuration settings and preferences.",
        "responses": {
          "200": {
            "description": "Get global config info",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Config" }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "global.config.update",
        "summary": "Update global configuration",
        "description": "Update global OpenCode configuration settings and preferences.",
        "responses": {
          "200": {
            "description": "Successfully updated global config",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Config" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Config" }
            }
          }
        }
      }
    },
...
    "/global/upgrade": {
      "post": {
        "operationId": "global.upgrade",
        "summary": "Upgrade opencode",
        "description": "Upgrade opencode to the specified version or latest if not specified.",
        "responses": {
          "200": {
            "description": "Upgrade result",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "success": { "type": "boolean", "const": true },
                        "version": { "type": "string" }
                      },
                      "required": ["success", "version"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "success": { "type": "boolean", "const": false },
                        "error": { "type": "string" }
                      },
                      "required": ["success", "error"]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "target": { "type": "string" } }
              }
            }
          }
        }
      }
    },
...
    "/project/current": {
      "get": {
        "operationId": "project.current",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Get current project",
        "description": "Retrieve the currently active project that OpenCode is working with.",
        "responses": {
          "200": {
            "description": "Current project information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        }
      }
    },
...
    "/session/{sessionID}/message/{messageID}": {
      "get": {
        "operationId": "session.message",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "workspace",
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "sessionID",
            "schema": { "type": "string", "pattern": "^ses.*" },
            "required": true
          },
          {
            "in": "path",
            "name": "messageID",
            "schema": { "type": "string", "pattern": "^msg.*" },
            "required": true
          }
        ],
        "summary": "Get message",
        "description": "Retrieve a specific message from a session by its message ID.",
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "info": { "$ref": "#/components/schemas/Message" },
                    "parts": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Part" }
                    }
                  },
                  "required": ["info", "parts"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NotFoundError" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "session.deleteMessage",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "workspace",
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "sessionID",
            "schema": { "type": "string", "pattern": "^ses.*" },
            "required": true
          },
          {
            "in": "path",
            "name": "messageID",
            "schema": { "type": "string", "pattern": "^msg.*" },
            "required": true
          }
        ],
        "summary": "Delete message",
        "description": "Permanently delete a specific message (and all of its parts) from a session. This does not revert any file changes that may have been made while processing the message.",
        "responses": {
          "200": {
            "description": "Successfully deleted message",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NotFoundError" }
              }
            }
          }
        }
      }
    },
...
    "/event": {
      "get": {
        "operationId": "event.subscribe",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Subscribe to events",
        "description": "Get events",
        "responses": {
          "200": {
            "description": "Event stream",
            "content": {
              "text/event-stream": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          }
        }
      }
    },
...
    "/tui/control/response": {
      "post": {
        "operationId": "tui.control.response",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Submit TUI response",
        "description": "Submit a response to the TUI request queue to complete a pending request.",
        "responses": {
          "200": {
            "description": "Response submitted successfully",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          }
        },
        "requestBody": { "content": { "application/json": { "schema": {} } } }
      }
    },
    "/instance/dispose": {
      "post": {
        "operationId": "instance.dispose",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Dispose instance",
        "description": "Clean up and dispose the current OpenCode instance, releasing all resources.",
        "responses": {
          "200": {
            "description": "Instance disposed",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          }
        }
      }
    },
...
    "/log": {
      "post": {
        "operationId": "app.log",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Write log",
        "description": "Write a log entry to the server logs with specified level and metadata.",
        "responses": {
          "200": {
            "description": "Log entry written successfully",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "service": {
                    "description": "Service name for the log entry",
                    "type": "string"
                  },
                  "level": {
                    "description": "Log level",
                    "type": "string",
                    "enum": ["debug", "info", "error", "warn"]
                  },
                  "message": { "description": "Log message", "type": "string" },
                  "extra": {
                    "description": "Additional metadata for the log entry",
                    "type": "object",
                    "propertyNames": { "type": "string" },
                    "additionalProperties": {}
                  }
                },
                "required": ["service", "level", "message"]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "worktree": { "type": "string" },
          "vcs": { "type": "string", "const": "git" },
          "name": { "type": "string" },
          "icon": {
            "type": "object",
            "properties": {
              "url": { "type": "string" },
              "override": { "type": "string" },
              "color": { "type": "string" }
            }
          },
          "commands": {
            "type": "object",
            "properties": {
              "start": {
                "description": "Startup script to run when creating a new workspace (worktree)",
                "type": "string"
              }
            }
          },
          "time": {
            "type": "object",
            "properties": {
              "created": { "type": "number" },
              "updated": { "type": "number" },
              "initialized": { "type": "number" }
            },
            "required": ["created", "updated"]
          },
          "sandboxes": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["id", "worktree", "time", "sandboxes"]
      },
...
      "OutputFormat": {
        "anyOf": [
          { "$ref": "#/components/schemas/OutputFormatText" },
          { "$ref": "#/components/schemas/OutputFormatJsonSchema" }
        ]
      },
...
      "UserMessage": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "pattern": "^msg.*" },
          "sessionID": { "type": "string", "pattern": "^ses.*" },
          "role": { "type": "string", "const": "user" },
          "time": {
            "type": "object",
            "properties": { "created": { "type": "number" } },
            "required": ["created"]
          },
          "format": { "$ref": "#/components/schemas/OutputFormat" },
          "summary": {
            "type": "object",
            "properties": {
              "title": { "type": "string" },
              "body": { "type": "string" },
              "diffs": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/FileDiff" }
              }
            },
            "required": ["diffs"]
          },
          "agent": { "type": "string" },
          "model": {
            "type": "object",
            "properties": {
              "providerID": { "type": "string" },
              "modelID": { "type": "string" }
            },
            "required": ["providerID", "modelID"]
          },
          "system": { "type": "string" },
          "tools": {
            "type": "object",
            "propertyNames": { "type": "string" },
            "additionalProperties": { "type": "boolean" }
          },
          "variant": { "type": "string" }
        },
        "required": ["id", "sessionID", "role", "time", "agent", "model"]
      },
...
      "Message": {
        "anyOf": [
          { "$ref": "#/components/schemas/UserMessage" },
          { "$ref": "#/components/schemas/AssistantMessage" }
        ]
      },
...
      "FilePart": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "pattern": "^prt.*" },
          "sessionID": { "type": "string", "pattern": "^ses.*" },
          "messageID": { "type": "string", "pattern": "^msg.*" },
          "type": { "type": "string", "const": "file" },
          "mime": { "type": "string" },
          "filename": { "type": "string" },
          "url": { "type": "string" },
          "source": { "$ref": "#/components/schemas/FilePartSource" }
        },
        "required": ["id", "sessionID", "messageID", "type", "mime", "url"]
      },
...
      "Part": {
        "anyOf": [
          { "$ref": "#/components/schemas/TextPart" },
          { "$ref": "#/components/schemas/SubtaskPart" },
          { "$ref": "#/components/schemas/ReasoningPart" },
          { "$ref": "#/components/schemas/FilePart" },
          { "$ref": "#/components/schemas/ToolPart" },
          { "$ref": "#/components/schemas/StepStartPart" },
          { "$ref": "#/components/schemas/StepFinishPart" },
          { "$ref": "#/components/schemas/SnapshotPart" },
          { "$ref": "#/components/schemas/PatchPart" },
          { "$ref": "#/components/schemas/AgentPart" },
          { "$ref": "#/components/schemas/RetryPart" },
          { "$ref": "#/components/schemas/CompactionPart" }
        ]
      },
...
      "BadRequestError": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "propertyNames": { "type": "string" },
              "additionalProperties": {}
            }
          },
          "success": { "type": "boolean", "const": false }
        },
        "required": ["data", "errors", "success"]
      },
...
      "NotFoundError": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "const": "NotFoundError" },
          "data": {
            "type": "object",
            "properties": { "message": { "type": "string" } },
            "required": ["message"]
          }
        },
        "required": ["name", "data"]
      },
...
    }
  }
}